refactor(opencode): separate usage accounting layers
